requirements.txt
setuptools~=50.3.0
discord.py~=1.4.1